feat(db): add Tag + NoteTagCrossRef entities, TagDao, migration 2→3
  - notes_tags junction uses snake_case columns (note_id, tag_id) with
    FK CASCADE on both sides; index on tag_id avoids full scans
  - Cross-ref writes (insertCrossRef, deleteCrossRefsForNote) live in
    NoteDao — keeps setNoteTags a single-DAO operation in NoteStoreImpl
  - Fix NoteStore.getNote / selectNote id: Int → Long (silent truncation
    bug on ids > Int.MAX_VALUE)

+++ androidApp/src/main/kotlin/com/jfranco/sharetosave/persistence/implementation/implementation/NoteStoreImpl.kt
@@ -35,7 +35,7 @@ class NoteStoreImpl @Inject constructor(
         note.copy(id = id)
     }
 
-    override suspend fun getNote(id: Int): Note? = withContext(ioDispatcher) {
+    override suspend fun getNote(id: Long): Note? = withContext(ioDispatcher) {
         noteDao.selectNote(id)?.toDomain()
     }
